Mbola mi-calcule Moyenne

+++ src/main/java/com/gestion/demo/repository/CV/CvRepo.java
@@ -3,7 +3,7 @@ package com.gestion.demo.repository.CV;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.gestion.demo.model.CV.Cv;
 
-public interface CvRepo extends JpaRepository<Cv, Long> {
+public interface CvRepo extends JpaRepository<Cv, Integer> {
 
 }
